Login - fix error message when user logs in with invalid credentials
The FailedAuthenticationController was previously using the errors
method from the ApplicationController; however, that is incorrect
since it does not inherit from it.  It purposely does not; otherwise,
the user has to be logged in for these actions to be execucted and
at this point, the user is not logged in.

Conflicts:

	src/app/controllers/failed_authentication_controller.rb