passwords are stored in secure format
Passwords are stored in hashed form only. The hash is calculated from the
password itself concatenated with ":" and with 64 long seed string and then
hashed 500 times using SHA512 algorighm. The seed string is different for
each user.

Each password is stored in the database in the form of hash + salt. Every
password is 192 characters long.

If passwords are stolen from Katello database attacker needs to spent giant
amount of effort to retrieve a single one. Thank to seed string and
repetition this will last centuries.