Roles - refactor self-roles to associated directly with a user
When a user (e.g. user1) is created, a 'self-role' is created for the
user (e.g. user1_role).  This self-role only applies to that user and
cannot shared/used by any other user.

Prior to this commit, the Role page would list all roles (including
self-roles).  With this commite, self-roles will now be omitted from the
Role page.  The self-roles have now been moved over to the User as part
of a "Roles & Permissions" tab.  On that tab, the roles for the user
may be assigned as well as permissions added/removed/updated from the
user.  Those permissions are essentially the 'self-role' permissions.