General UI - fixes in before_filters
While working on activation keys in remote branch, I noticed a couple of
issues with some of our before_filters.  First, many were fetching data
from ActiveRecord (e.g. find), but not catching exceptions.  As a result,
the error handling behavior of the filter was not being executed.  Second,
even if the exception was caught the error notice would not be displayed
to the user, due to the fact that the notices are handled by an after_filter
(flash_to_headers); however, the render/redirect halts the filter chain.

This commit is to address those issues within the controllers.