Views - update grid in various partial to account for panel size change
This is minor change to some of the view partials to take advantage of the
additional space that has been made available in the right pane due to
resizing of the left panel.

Note: activation keys partials will be updated in the a-keys branch
to avoid merge conflicts with work currently being done there.