Activation Keys - system template support on crud
This commit is a new implementation for the system template support.  The primary
changes are in editing an activation key.  Now, when a user clicks environment, they
will be presented with a jquery ui dialog to allow them to select the environment.
If they choose an env, they will also be presented with a list of system templates
available to that environment.