719932 - subscription upload - fix endless spinner and error handling
This commit addresses the issue where after a subscription upload
via the UI (successful or unsuccessful) the spinners continue after
the action has completed.  This is an issue of the ajaxComplete not
being triggered after the action and may be related to file uploads
not truely being treated as ajax requests.  In order to address this,
we are going to trigger ajaxComplete manually.

This commit also addresses an issue where on subscription upload error
where we would attempt to render the subscriptions partial without valid data
causing an ugly haml error to be shown.  In order to address this, we
are purposely rendering "nothing", so that the user may select a different
manifest, if desired, for upload.