Notices - update js to check for existence of xhr in response
While working on subscription file upload, observed a js error in notices.
The issue being encountered is that we were looking for notices details
in the file upload response; however, since that request is not a true
ajax request, there was no detail to be retrieved.  That said, notices
was not checking to see if the variable was defined before using it.