730738 - fix notification on invalid login attempt
Removing the UserSessions::invalid from the invalid login attempt flow.

Prior to this change, the flow was similar to:
 POST user_session -> redirect_to invalid_session_url -> redirect_to new_user_session_url

After this change, the flow is:
 POST user_session -> redirect_to new_user_session_url

This removes the additional redirect in the middle of the flow.

This change also moves the flash/notice to the failed auth controller and
uses flash direct vs AppController::errors.  Since this controller does
not inherit from AppController (and it should not), it cannot use that method.
That is not, however, a problem since we simply want to display a notice and
not persist it.