Notices - fix to enable login/logout notices to appear
The login/logout notices were being generated, but not displayed.
This was used to the flash.now which expects the notice to be displayed
immediately as part of the same request; however, these actions are both
redirects, so notice should be displayed on next request.

In order to address this, removed the '.now'.

In additional, removed the check to see if the response is for
an ajax request.  This is not really necessary, since we are ultimately
handling the notices the same for all cases, except redirects which
we want to display on next request.