Changeset History - improve grid usage on edit view
With the recent changes to the left panel sizing for changesets
(i.e. to use 1 column), the format of the right edit pane runs
together a bit.

This commit is to change some of the grid usage to take advantage
of the increased size of the right pane.

In a future commit, we'll address the sizing of the jeditable
text field and area; however, wanted to get this initial change in.