Role - Activation Keys - add resource type, model and controller controls
This is the initial commit to add the new roles/permissions logic to the
activation key model and controller.

For activation keys, we are modeling sync plans in that they will be defined
as a separate resource type with verbs of read_all and manage_all.  This
means that when a user defines permissions for akeys, they will not be specifying
the individual tags that a user has access to.