auto_tab_index - introduce a view helper to simplify adding tabindex to forms
This commit is to add a view helper called auto_tab_index.  It is
intended to simplify adding a tabindex to a form.  Rather than state
:tabindex => 1, :tabindex => 2, :tabindex => 3... use :tabindex => auto_tab_index.
The helper will take care of assigning the next value.  This makes it easier
to add new fields to the form w/o having to update all of the indexes.