Changes to model id's in views.
This is the first in a series of commits aimed at changing the way model
and resource ids are handled in the HTML and javascript layers.  The
goal is to scope to each id by the model or resource instead of just the
object instance id alone.  For example, a role will now be identified in
the HTML by "role_3".  This commit changes this behavior specifically
for roles while keeping the behavior consistent for all other resources.