Views - update to use image_tag helper
This commit is to convert all view usage of %img to use the
RAILs image_tag helper.

While the current approach works fine when in development env using
webrick, it won't work when we support apache/thin and have a
application prefix of /katello.

Before this change, using the img definition, the html would
be generated using src="/image/...".  When we use image_tag, RAILs
will take care of adding our application prefix, which would result
in src="/katello/image..."