app server - setting relative paths on fonts/images in css & js
Prior to this change, the inclusion of assets such as images and fonts
from within css/scss/js used a fixed path (e.g. /images). This worked
fine when deploying the app using webrick and using / as a base path
for the app (e.g. http://0.0.0.0:3000/); however, it does not work
as we introduce thin+apache and use a base path of katello for the app
(e.g. http://0.0.0.0/katello).

This commit is to update those assets to use a relative path for the
location of the images.   This relative path will need to be based
on where the files will reside when the app is running (e.g. public/...).