Views/JS - updates to support /katello prefix
This contains the changes to address the hardcoding of resource
URLs within the javascript files.  Where possible, have updated
the view haml to use data-url.  There are, however, cases where
this cannot be easily done.  In order to address those cases,
added KT.common.rootURL, so that any JS can easily obtain the
rootURL (e.g. /katello) and add it to the resources it is using.

Note: KT.common.rootURL will avoid impacting the various JS, if
we decide to change the project name or prefix in the future.

Note: there will be more commits coming to address any resource
hardcoding within the haml and helpers...