script/rails - adding back in... needed to run rails console
As part of the thin migration we removed script/rails.  The reason being
developers should start the application using script/thin instead.  Today,
we found that "rails console" is no longer working.  It turns out that in
order for that to work, script/rails must exist.

I am re-adding the original script/rails back in.  This script includes
both the standard logic that is provided by "rails new katello" as well
as the custom code we had placed in it for handling SSL when using webrick
and looking at version info for things like candlepin.

It should be noted, that we still do not intend for developers to run
the app using script/rails (Please use script/thin), but if they need
to access console, "script/rails console" is available.