app server - updates to use thin Rack handler vs script/thin
Prior to this change, we found that developers using rubymine could
not really run the app within the IDE using the thin app server
and applying the /katello prefix.  This change is to address that issue,
but it applies a change to how a developer will start katello in a dev
env.

A few options that exist are:

Scenario 1 (recommended): Start using prefix and thin:

  RAILS_RELATIVE_URL_ROOT=/katello script/rails server thin

Scenario 2: Start using thin w/o prefix:

  script/rails server thin

Scenario 3: Start using webrick:

  script/rails server

  Note: for scenario 3, if user specifies "use_ssl: true" in the
  katello.yml, the service will be accessed with https; otherwise,
  it is http.

For thin, SSL direct access is not supported by the above scenarios.
If SSL is required with thin config, it is recommended to use an Apache front-end
similar to the production deployment.  The reason being the RAILS Rack
Handler for Thin doesn't currently provide mechanism to specify
the SSL details (e.g. enable, key, cert, verify...)