search - systems - update to properly handle autocomplete
This commit contains changes to improve the behavior of search auto-complete
for systems (All and By Environment).  If on 'All' systems page, auto-complete
will filter all systems; however, if on the page w/ the environment selector
it will only filter on those in the current selected env.

With this commit, if a page has the environment selector, the developer
can basically do the following and auto-select should filter on env:
1. add a 'by_env' scope to the model (e.g. system.rb)
2. add the env to the panel options in the controller action that
   lists the elements (e.g. index or environments in this case): e.g.
  @panel_options[:search_env] = @environment.id

can basically add a 'by_env' scope to the model, add