notices - change to support closing previous failure notices on a success
This commit is to provide a change in behavior for notices.

Today, when error notices are created, they are created as sticky and require
user to close them manually.  This is a bit annoying if the notice is for a
validation error (e.g. creating Provider) and then the user successfully performs
the same action again.  In that case, user gets success notice, but must still
manually close the error.

With this commit, katello will attempt to close failure notices that are
associated with the same request type as a later success.  It will identify
request type using the 'controller' and 'action' that is being requested.

For example, if user attempts to update a provider (controller: providers,
action: update)
- if the first attempt is a failure, error notice created
- if the second attempt is a success, error notice cleared, success notice created

In the above scenario, we purposely do not look at the lowest level which
might be: 'provider X, field Y updated'.  Doing so, would require storing a lot
more context and have much more impact on the views...etc.