a-keys - update setting of env and system template on general tab...
This is the first of several commits to come to change the UI for activation
keys.

This commit changes how the env and system template are set on the edit
screen.  Based on feedback from UXD, we will no longer use the jquery
UI dialog for changing the values of env and template.  Instead, we'll
use the 'environment selector' (similar to what is used on pages like
promotions, systems, changeset history..etc).  In order to use the
env selector, did make several changes to support enabling it to be
used within a pane.