a-keys - initial changes for applied subscriptions page
This commit contains the initial code for supporting the
activation key applied subscriptions page/tab.  Additional changes
will be needed to incorporate product family (or whatever we decide
to use), virtual/physical machines...etc

Note: in this commit, also removing the former code for subscriptions.
In doing so, am repurposing the _subscriptions partial for code that
is common to applied/available subscriptions.