741444/741648/739981/739655 - update *.js.haml to use the new KT namespace for javascript
This commit addresses the 4 bzs above.  Recently, a KT namepace was introduced and the panel
was included as part of it.  All of the *.js files were updated; however, we have a few additional
changes to the *.js.haml files.  These are files that the controllers can render and triggers the
js to be executed.