a-keys - js fix so that clearing filter does not leave children shown
Without this change, when the user enters a filter, all entries
in the subscription table that match it are shown.  This is good and
expected.  But, when the filter is cleared, any rows that are part
of a 'collapsed' parent, remain visible.  This should not be the case.
As a result, this commit will hide those children.