Move installation-related parts to */install.pp.
Hello,

please apply the following patch to katello master. It moves all the
installation-related steps to */install.pp while keeping the
configuration / setup in place.

The current katello-configure script which does

	echo include katello | puppet ...

will then only be working after all packages were already installed,
ideally via

	yum install katello-all

which is the ultimate goal. Please note that the patch

	[PATCH] Upon the db:seed, we also have to db:migrate.

is needed for this approach to work correctly, so please apply that
one as well. Once the patches are committed and packages built and
in yum repo, we will need to update the wiki to read

	rpm -Uvh http://path-to-katello-repos.rpm
	yum install -y katello-all
	katello-configure

If for some reason the user wants to also use puppet to install
the packages, they would do

	echo include katello::install | puppet ...

instead of

	echo include katello | puppet ...

I recommend not to document this approach on the wiki and only leave
it as the last resort way for situations where the rpm-based
installation really cannot be used, for people who know what they are
doing.