739680 - include candlepin error text in error notice on manifest upload error
With this commit, if the user attempts to upload a manifest and it generates
an error from candlepin, the error text (displayMessage) from candelpin will
be included in the notice generated to the user.

E.g.
Subscription manifest upload for provider 'Red Hat' failed.
Reason: Import is older than existing data