panel - minor update to escape special characters in id
The promotions page uses the errata ID as an html id.  The issue
encountered with this is that the errata ID is similar to "RHSA-2011:0310",
which contains a character that is special to jquery (i.e. ":").  This
change will run the escapeId on the selector so that errata elements
are properly identified.