Add support for answer files.
Hello,

please review a patch which is a start of moving default values from
the params.pp files to an answer file, to be able to then merge
default and custom answer files, plus command line options to
katello-configure. With the default values in this particular patch,
it's possible to do

	PGPASSWORD=katellopw psql -h localhost -U katellouser katelloschema

so the values are properly processed and propagated.

I'll leave it up to you guys to decide which values should be
configurable via answer files / CLI options -- those should then be
specified in the default-answer-file. Please add comments (use hash
signs at start of lines) before each of the default value because
another patch will follow which will generate a man page from the
default-answer-file and the comments will be put into the man page.

Thank you.