741656 - roles - search - chgs for search by perm type and verbs
This commit is to address a couple of issues with searching for
a role based on permission type and verb.

It also removes the search by tag.  Since the tag model is no
longer active record, we cannot use it in the search (since scoped_search
is based on active record).  Will look for alternatives to support search
by tag if we find it is a useful search.