744067 - Promotions - Errata UI - clean up format on Packages tab
This commit is to clean up format of the packages tab.  Previously, it
just dumped the pkg list to the pane.  This change will put it in tabular
format with a filter to simplify using the data presented.  This commit
only includes the name of the package; however, we may want to add additional
details in a future commit.  Attributes available per package include:
name, version, release, arch, epoc, src rpm, filename rpm and sum.

This commit also updates the view to use the tupane_layout.