Errata Filters - ui - initial changes to promotions breadcrumb
This commit contains the changes to the promotions breadcrumb to support
the errata filters.  This includes retrieving all errata, or by category or type.

Note: this does not include changes in the controller to actually return the
correct errata.  There is work ongoing in the CLI to support this as well;
therefore, did not want to duplicate work.