Users - add email address (model/controller/view)
This commit is to add the email address as a required element when
a user is created.  It also allows for editing of the address
after the user is created.

Note: email address will be necessary in order to be able to
send the user a 'reset' password, upon request.