password reset - update so that emails are sent asynchronously
This commit includes changes to send the password reset and login
emails asynchronously.

There was 1 small change needed to AsyncOperations so that the 'deliver'
would be invoked, if the operation is on a Mailer.  While this seems
a bit odd, the DelayedJobs gem does something similar (i.e. invokes deliver).