password reset - fix issue w/ redirect to login after reset
Prior to this commit, when user would change password using password
reset screen, they would be redirected to the login screen.  While this
looked appropriate at first glance, after the user logged in, they would
be sent to the dashboard, but the resource url in the browser was
that of the password reset request.

In order to address this, we are removing the redirect from the
controller and performing the redirect within the js itself.