installed-products - API supports consumer installedProducts
The installedProducts attribute send subscription-manager to the server while
registering the system. When something changes in locally installed products,
also rhsmcertd sends this information.

Note: to force rhsmcertd to update the system info it's good to restart the
service. OTOH to really send this information a change in /etc/pki/products
(adding or removing products) has to be made.