752863 - katello service will return "OK" on error
I have modified your init script and enhanced "status" call - now it
properly checks for all PID files.

On top of that I have found we are running thin with a daemon command - this
is not necessary since thin already daemonize processes. I removed that.

I also removed unnecessary check_permission call because thin reports error
when pid is not writable. This test was because of previous server
(Webrick).

Please note this does NOT fix the bug, it's an improvement. We need to do
more investigation.