packages - backend changes to support package and package group install/uninstall
This commit contains some backend changes to support package and package group
install and uninstall for a system.  From a pulp perspective, these actions
are asynchronous.  For Katello, with this commit we'll schedule them as
delayed jobs and use the pulp wait for task, to wait for them to complete.
This is being done so that we can use this capability to show the status
of the action in the UI.   That said, we'll need to do some testing to see
if this is a good approach or if we need a more general solution to check
for the status of package/group actions and avoid blocking within delayed jobs.