packages - make pkg/group install/remove scheduling actions synchronous
This commit makes the actions for scheduling pkg and group installs/removes
synchronous.   From a pulp-perspective, the execution of the actions will
be asynchronous; however, we do not need to schedule them asynchronously... :)
In the future, we'll be adding changes to take the task received as part of
scheduling, to monitor it for status/completion.