packages - update ui to show that action is in progress on pkg or group
After the user initiates a package install/update/remove or package group
install/remove, update the UI to show a spinner and status indicator
showing the action that is in progress.  There will be additional commits
for updating that status after the initial scheduling; however, those
are dependent on some model changes coming.