reimport-manifest - don't delete untracked products when importing
As a work-around for deleting an organization with some product not tracked in
Katello this products were deleted right after the manifest. The problem with
with this showed up, when an uebercert was created (and therefore a product that
is not in the Katello). When reimporting, Katello tried to remove this ueber
product which caused validation error on the Candlepin side.

Therefore now we don't delete any product on manifest import. It seems this
problem with untracked products that caused problem on deleting an organization
got away with new manifest.