packages - update so that loading System->Packages shows actions in progress
Prior to this commit, when rendering the packages pane, we only showed
the list of packages in the system's profile.

With this commit, we'll look to see if there are any pending package
actions for the system.  If so, we'll display those and continue to monitor
them while the pane is open.