758620 - promotion - fix ability to promote a distro
There was a recent change to katello to store details of a
repo within the katello model vs having all info come from pulp.
These changes made resulted in repo.id becoming an internal
katello id and the addition of repo.pulp_id.

This commit is to make the promotion of distro work w/ the new model.