system packages - generate validation error if action already in progress
Prior to this commit, if user attempted to install/remove a pkg or group
(via the content box) and one was already inprogress for that pkg or group,
we would generate an alert.

This commit will refactor the validation error logic recently added
for validating the format of the list entered... and reuse it for displaying
this validation error using consistent behavior.