768047 - promotions - let user know if promotion fails
Prior to this commit, if a promotion fails, the promotion page
continually indicates 'Promoting...'.  As a result, the only way
for the user to know that it failed was to reload the page.

With this commit, the UI will update the status to indicate
'Promotion failed'.  In addition, it will allow the user to resubmit
the promotion request.

This commit does not show details on the actual error that occured.
Will be looking at adding more details in via a 'notice'; however,
notices are owned by the ApplicationController and the
promotion is handled as an async delayed job, which does not have
access to the controller.  As a result, will have to do a bit of
refactoring to enable that access.