moving /distributions API into /repositories path
Since Katello does not have any distribution model class and Pulp is not
able to provide what repository this distribution belongs to, we would need
to change the route to:

/repositories/:repository_id/distributions/:id

So the controller code would be able to fetch repository/product and check
permissions.

On top of that, cli command distribution info now require new parameter
"repo_id".