fix broken tests resulting from RAILS / katello name collision
This small commit changes the name of the Glue::Pulp::Repos instance
method from repository_url to repo_url.  We recently added a repositories
resource to routes.rb that resulted in a name collision with the beforementioned
method and the repository_url method that RAILS provides by default.  We
could have changed the route; however, we use repositories elsewhere in
the routes.rb for other interfaces and we use repo_* elsewhere in
Glue::Pulp::Repos; therefore, this changed seemed to keep the interfaces
the most consistent.