768047 - notices - update to enable usage in controller or model
This is a second commit related to bug 768047.  For that bug we need
to be able to display error notices related to promotions of changesets.
The challenge, however, is that prior to this commit notices were part
of the application controller, but changeset promotion is handled as
a delayed (async) job.  As a result, a notice could not be easily created.
This commit addresses that restriction.

This commit pulls the logic for generation of notices out of the
application controller and in to a common module (lib/utils/notices.rb).
That module is included in the application controller for access by
all controllers inheriting from it and it may also be included
in models that require the ability to generate a notice (e.g. changeset.rb).

It is important to note that any notice generated as part of an
aysnc job, must be created as an async notice. For example,
  notice notice_text, {:synchronous_request => false}

As part of these changes, we also removed the the "errors" method
for generating error notices.  This method name conflicts with ActiveRecord
models.  The following is an example:

  Before:
    errors error_text

  After:
    notice error_text, {:level => :error}