gpg-cli - use immutable value gpgurl in CP
Prior this commit, the path used to retrieve GPG content might change after
content was created ("/api/gpg_keys/<key_id>/content").

After this modification, the url is
("/api/repositories/<repo_id>/gpg_key_content"). The content of this url
changes based on the currently assigned gpg_key. That means that after changing
gpg key assigned to the repo only the content changes but the url stais the
same. So that there is no need to update the CP content somehow.