750354 - activation key - do not highlight to change template when there are none
When a user creates or edits an activation key, if they select an
environment that has one or more system templates, highlight to the user
to change the template.  If the new environment does not have any
templates to be selected, there is no need to highlight the select
box.