perms - fake /api/packages/ path for rhsm
After registering, subscription-manager does the get on /katello/api after
which the rhsm.log shows:

012-01-18 08:05:03,165 [INFO]  @cache.py:209 - Server does not support
packages, skipping profile upload.

This appears to occur because previously packages was a top level API object
and thus had an entry in the hash sent by the /katello/api request. Now that
packages is nested under repositories, it no longer appears in the hash
causing subscription-manager to look for the existence of a "packages" entry
in line 209 of cache.py and spits out the above error if it cannot find it
and thus never uploads the package +profile.

Will create a "fake" /packages path that won't be really there just to
satisfy it. Packages are uploaded using the /systems/ path.