782518 - providers - removed org from resource to fix search history
Before this commit, accessing the providers page would be via a
resource like katello/organizations/[:id]/providers.  This is inconsistent
with the other pages in the UI.  In addition, it results in not being able
to properly store/access the history for searches performed on this page.

As a result, updating this page to be consistent with the rest, resulting
in using the following resource:
  /katello/providers