783329 - fix indexing of pkgs for elastic search and use in system templates UI
This commit makes some modifications to the indexing of packages
for elastic search.  The primary change is to allow us to support
'autocomplete' by providing an ngram filter on packages.  This also
allows for us to use the same name_search when the user provides the
full package name or partial.

Also, updated the system template UI to use the elastic search query
vs pulp.

Note: for system templates the auto-complete will sort the results based
on 'most relevent'... We may want to improve this in the future to only
show those that are absolutely relevant (e.g. only show entries that contain
exactly the characters typed.)