748060 - fix bbq on promotions page
Due to a race condition, the bbq for the promotions page was not
behaving properly when user was viewing an item from the content
tree (e.g. distro, pkg, errata).  Essentially, the panel would
receive and process the hash_change event for the list of items
(e.g. distros) was rendered in the content tree.  As a result, it
would not be able to display the item that was previously selected
(e.g. a distro).