786222 - system templates - allow selecting of distro based on repos
Prior to this commit, when the user created a system template, if
they added a product to a template (and that product contained at least
one repo that had a distribution associated with it), the user could
associate a distribution with the template.  However, if the user
instead added the repository to the template (instead of the product),
the user could not associate a distribution with the template.  I.e.
there had to be a product.

With this commit, if the template has a product and/or repo associated
with it and they have a distro, the user can select it.