771886 - system packages - fix ui staying in processing state on pkg install
This commit is to address an issue found with system package/group install
and removes.  In firefox 3.6, it was possible for the UI to remain in
'processing' state after an install/remove action was initiated.  This
issue was actually due to an error in the javascript where 'for' loop
initializer variables were not explicitly defined, resulting in
a js error an the logic ignoring the status updates.