771957 - let repositories deal with content deletion
Only marketing products must deal with deleting product - content assignment in
CP. For engineering products there are repositories created, which delete the
content when they are removed. This change prevents trying to delete the
content twice while deleting product.

The other posibility would be to prevent deleting of conent in repository in
case product is being deleted and let product handle it all, but there would be
necessary a mechanism for finding out if the repo orchestration happens during
product orchestration or not - when not repo should handle content deletion
itself.