786598 - system templates - improve support for repos having same name
In Katello, it is possible to have multiple repositories with the same
name (e.g. repoX) across multiple products.

This commit contains changes to improve Katello's ability to support
that with System Templates.  Essentially, allow the user to pick a
specific repo and add it to a template.  This applies when selecting
the repo from the left content tree or when using the auto-complete
search box from within the right template tree.

In the case of auto-complete, as the user types the repo name,
Katello will provide the repos that match along with the product
they are part of.  If the user picks from the auto-complete results,
only the selected repo will be added to the template.  If the user
ignores the auto-complete results and adds the repo explicitly
based on what they typed, then Katello will add all repos that
match to the template.