activation keys - remove 'allocated' from model and code
This commit removes 'allocated' from the key_pools.  In the initial
implementation for activation keys, we allowed the user to assign
a quantity (allocation) of a given subscription to the key.  After
discussion, we decided to revamp it to no longer associate quantity
and instead have the keys used with 'auto-subscribe'. This approach
eliminated the need to have several different keys to meet the needs
of consumers (e.g. 2 socket, 4 socket...etc).