789456 - pre-save and post-save queues in orchestration
Rename current one queue processed before saving a record to pre_queue and add
post_queue processed after saving.

This is required in case we need to use id generated in Katello db in
orchestration (e.g. make a refference in CP to object in Katello - environments
orchestration).

When a task on post_queue fails, also completed tasks in pre_queue are
processed.