741499-Added code to deal with weird user current org behaviour
1) User has access to Orgs A,B, C and C is its current org.  Admin revokes
access to org C from the user. User when he clicks on any page will
automatically be logged out with a message stating

"""
You current organization is no longer valid. It is possible that either the
organization has been deleted or your permissions revoked, please log back in
to continue.
"""

2) User has access to no orgs. Admin then provides access to orgs A, B. The
user should immediately get assigned to org A just as he would reload or click
any page. (it is as if the user relogged in.)

Also includes many unit test fixes