789516 - Promotions - fix ability to add products and distros to a changeset
A recent change to the promotions logic requires the data-product_id to
be an array of IDs.  If it isn't then, the javascript will not properly
add the element to the changeset.  This change is to address this for
products and distributions.  Other elements (e.g. pkgs, repos, errata)
are already set correctly.