786179 - notices - delete all displaying notices after completed
Before this commit, it was possible that user would click 'delete all'
to destroy notices and possibly see 'some' (e.g. handfull) notices still
rendered.  This is due to a race condition between the destroying of
the notices (in DB+search server) and the re-rendering of the notices.

In order to address this, we are converting the page to use ajax
(consistent with the rest of the UI) and removing the content from
the table vs asking the server for 'an updated list'.