782518 - search history - parse out host using URI class vs using HTTP_HOST
Search history is implemented to look in to the HTTP_REFERER to
determine which page the user is requesting search history for.
The HTTP_REFERER, however, contains the full URI (e.g.
http://this.server.com/katello/organizations).  For this to work, we
need to strip anything prior to the katello prefix.  Previously,
we would do this by looking at the HTTP_HOST in the incoming
request.  The issue with this is that in production configuration,
we see the host no longer matches what is in the referer.  This is
likely due to the apache front end.  In order to address this,
we can use the ruby URI class to parse out the path.