796852, 789533 - search - update to handle - search queries
This commit is to address the issue raised by the above bugs with
respect to inability to search for repo and package names that contain
a -.  That said, this commit is a broader fix to handle searching
any content that contains a -.

The search query syntax used by Katello is based on Lucene query syntax.
(e.g. http://lucene.apache.org/core/old_versioned_docs/versions/3_5_0/queryparsersyntax.html).
In that syntax, the - is a predefined operator.  In
order to not have that behavior applied, the user must escape the
operator (e.g. /-).  Unfortunately, that is a hassle for Katello users since
- is commonly used as a seperator in things like package and repo names.

As a result, we have chosen to disable the - operator in queries from
Katello.  This means that any - entered will be automatically escaped.

As an alternative, we could have chosen not to use Lucene syntax at all;
however, Lucene does provide several of the capabilities that are useful
to uses (e.g. such as support for AND, OR...etc).