786572 - elasticsearch - reduce config to use 3 shards vs default (5)
Elasticsearch (by default) is configured to use 5 shards and 1 replica.
That allows the server to grow scaling across up to 10 servers.  While
that is nice, it does come at the cost of some memory.  We could reduce
the number of shards to 1, given that our initial deployment configuration
is only 1 node; however, that would not facilitate future growth/scalability.
In order to strike a balance between scalability and minimize impact to memory,
we are reducing the configuration to use 3 shards.

Note: from some basic testing with 5 vs 3 shards, observed the following:
- 3 shards : VIRT(~1650m), RES(~395m)
- 5 shards : VIRT(~1720m), RES(~452m)

Scenario run in both configurations before arriving at the above was:
- define 2 providers (1 w/ rhel6 & 1 w/ rhel6.1)
- sync both 2 times
- promote both 2 times